refactor: Expose the pact base url as an env variable relatesto: form3tech/project-euro-board#1291
We are working towards removing viper from the sepadd-gateway code
and eventually from all gateways. This pact testing library uses
viper as a means of providing the stub url (with randomly selected
port) to the gateway pact tests. If we want to remove viper from the
gateway, we first need to remove this dependency on viper in pact
testing.